# 3° Anno/Programmazione per il Web/Progetto/WASAText/doc/api.yaml
openapi: 3.0.0
servers:
  - url: "http://localhost:3000"
info:
  title: WASAText
  description: una piattaforma di messaggistica che supporta chat individuali e di gruppo, accessibile tramite PC
  version: 1.0.0
...
paths:
...
  /chats/{chat_id}/messages:
    parameters:
        - name: chat_id
          in: path
          description: ID univoco della chat.
          required: true
          schema:
              type: string
              minLength: 1
              maxLength: 64
              pattern: '^[a-zA-Z0-9_-]{1,64}$'
    get:
      tags:
        - messaggi
      summary: Ottieni tutti i messaggi di una chat ordinati in ordine cronologico decrescente
      description: Restituisce la lista di tutti i messaggi appartenenti alla chat specificata, ordinati dal più recente al meno recente.
      operationId: getMessages
      parameters:
        - $ref: '#/components/parameters/Ordinamento'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Elenco messaggi recuperato con successo
          content:
            application/json:
              schema:
                type: array
                description: Lista di messaggi della chat
                items:
                  $ref: '#/components/schemas/Message'
                minItems: 0
                maxItems: 1000
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - messaggi
      summary: Crea e invia un messaggio in chat 
      description: |-
        Permette di creare e inviare un nuovo messaggio all'interno della chat specificata.
        Restituisce l'oggetto messaggio creato.
      operationId: sendMessage
      security:
        - bearerAuth: []
      requestBody:
        description: Dettagli del messaggio da inviare
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMessageRequest'

          # Nuovo supporto per l'invio di messaggi con allegati
          multipart/form-data:
            schema:
              type: object
              description: Oggetto per l'invio di file multimediali
              properties:
                content:
                  type: string
                  description: Testo opzionale del messaggio
                  minLength: 0
                  maxLength: 1000
                  pattern: '^[\s\S]*$'
                type:
                  type: string
                  description: Tipo di contenuto inviato
                  enum: [text, image]
                  default: text
                file:
                  type: string
                  format: binary
                  description: File immagine da caricare
                  minLength: 1
                  maxLength: 1073741824 # 1 GB
  
      responses:
        '201':
          description: Messaggio inviato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'


  # SEZIONE 3.1 - Eliminazione di un messaggio specifico
  /chats/{chat_id}/messages/{message_id}:
    parameters:
        - name: chat_id
          in: path
          description: ID univoco della chat.
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 64
            pattern: '^[a-zA-Z0-9_-]{1,64}$'
        - name: message_id
          in: path
          description: ID univoco del messaggio.
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 64
            pattern: '^[a-zA-Z0-9_-]{1,64}$'
    delete:
      tags:
        - messaggi
      summary: Elimina un messaggio specifico dalla chat
      description: |-
        Permette di eliminare un proprio messaggio specifico all'interno della chat.
      operationId: deleteMessage
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Messaggio eliminato con successo
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # SEZIONE 3.2 - Risposta ad un messaggio esistente
  /chats/{chat_id}/messages/{message_id}/reply:
    parameters:
        - name: chat_id
          in: path
          description: ID univoco della chat.
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 64
            pattern: '^[a-zA-Z0-9_-]{1,64}$'
        - name: message_id
          in: path
          description: ID univoco del messaggio.
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 64
            pattern: '^[a-zA-Z0-9_-]{1,64}$'
    post:
      tags:
        - messaggi
      summary: Rispondi a un messaggio esistente nella chat
      description: |-
        Permette di rispondere a un messaggio esistente all'interno della chat specificata.
        Restituisce l'oggetto messaggio creato come risposta.
      operationId: replyToMessage
      security:
        - bearerAuth: []
      requestBody:
        description: Dettagli del messaggio di risposta da inviare
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMessageRequest'
      responses:
        '201':
          description: Risposta al messaggio inviata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'


  # SEZIONE 3.3 - Reazione ad un messaggio specifico
...
  /chats/{chat_id}/messages/{message_id}/forward:
    parameters:
        - name: chat_id
          in: path
          description: ID univoco della chat.
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 64
            pattern: '^[a-zA-Z0-9_-]{1,64}$'
        - name: message_id
          in: path
          description: ID univoco del messaggio.
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 64
            pattern: '^[a-zA-Z0-9_-]{1,64}$'
    post:
      tags:
        - messaggi
      summary: Inoltra un messaggio a un'altra chat
      description: |-
        Permette di inoltrare un messaggio esistente verso una o più chat di destinazione.
        Crea un nuovo messaggio per ciascuna chat di destinazione, con un riferimento al messaggio originale.
      operationId: forwardMessage
      security:
        - bearerAuth: []
      requestBody:
        description: Elenco delle chat di destinazione per l'inoltro
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForwardMessageRequest'
      responses:
        '201':
          description: Messaggio inoltrato con successo
          content:
            application/json:
              schema:
                type: array
                description: Lista di messaggi inoltrati
                items:
                  $ref: '#/components/schemas/Message'
                minItems: 0
                maxItems: 1000
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

# ------------------------------- COMPONENTS ---------------------------------- #
components:
...
  schemas:
...
    Message:
      type: object
      description: Rappresenta un messaggio all'interno di una chat
      required:
        - id
        - chat_id
        - sender_id
        - type
        - timestamp
        - status
      properties:
        id:
          type: string
          minLength: 6
          maxLength: 64
          pattern: '^[a-zA-Z0-9_-]{6,64}$'
          example: "messaggio1"
          description: Identificatore univoco del messaggio
        chat_id:
          type: string
          minLength: 6
          maxLength: 64
          pattern: '^[a-zA-Z0-9_-]{6,64}$'
          example: "chat123456"
          description: Identificatore univoco della chat
        sender_id:
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^[a-zA-Z0-9_-]{1,64}$'
          example: "utente789"
          description: Identificatore univoco dell'utente che ha inviato il messaggio
        type:
          description: Tipo di messaggio (testo, immagine, file, audio)
          type: string
          enum:
            - text
            - image
            - file
            - audio
          example: text
        content:
          type: string
          minLength: 1
          maxLength: 1000
          description: Testo del messaggio (se type = text)
          pattern: '^[\s\S]{1,1000}$'
          example: "Ciao, come stai?"
        media_url:
          type: string
          format: uri
          description: URL del file multimediale (solo se type ≠ text)
          minLength: 1
          maxLength: 2048
          example: "http://example.com/uploads/photo123.png"
          pattern: '^https?://.+$'
        timestamp:
          type: string
          format: date-time
          example: "2025-10-30T14:48:00.000Z"
          description: Data e ora di invio del messaggio
        status:
          description: Stato del messaggio (inviato, consegnato, letto)
          type: string
          enum: 
            - delivered
            - read
        reactions:
          type: array
          description: Lista delle reazioni a questo messaggio
          items:
            $ref: '#/components/schemas/Reaction'
          default: []
          minItems: 0
          maxItems: 50

    # Creo uno schema per gestire gli snippet di messaggi 
...
    CreateMessageRequest:
      type: object
      description: Rappresenta la richiesta di creazione di un nuovo messaggio
      required:
        - type
      properties:
        type:
          type: string
          enum:
            - text
            - image
            - file
            - audio
          description: Tipo di messaggio
        content:
          type: string
          minLength: 1
          maxLength: 1000
          pattern: '^[\s\S]{1,1000}$'
          example: "Ciao, come stai?"
          description: Testo del messaggio (se type = text)
        media_url:
          type: string
          format: uri
          minLength: 1
          maxLength: 500
          description: URL del file multimediale (solo se type ≠ text)
          example: "http://example.com/uploads/photo123.png"
          pattern: '^https?://.+$'
        reply_to_message_id:
          type: string
          minLength: 6
          maxLength: 64
          description: Identificatore del messaggio a cui si sta rispondendo (opzionale)
          pattern: '^[a-zA-Z0-9_-]{6,64}$'
          example: "messaggio5678"

    # Creo uno schema per le reazioni ai messaggi (tramite emoji)
    Reaction:
      type: object 
      description: Rappresenta una reazione a un messaggio tramite emoji
      required:
        - id
        - message_id
        - user_id
        - emoji_code
        - created_at
      properties:
        id:
          type: string
          minLength: 6
          maxLength: 64
          pattern: '^[a-zA-Z0-9_-]{6,64}$'
          example: reaction1
          description: Identificatore univoco della reazione
        message_id:
          type: string
          minLength: 6
          maxLength: 64 
          pattern: '^[a-zA-Z0-9_-]{6,64}$'
          example: message1234
          description: Identificatore del messaggio a cui è associata la reazione
        user_id: 
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^[a-zA-Z0-9_-]{1,64}$'
          example: user1
          description: Identificatore dell'utente che ha inserito la reazione
        user_name:
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^[a-zA-Z0-9._-]{1,64}$'
          example: Maria
          description: Nome dell'utente che ha inserito la reazione
        emoji_code:
          type: string
          minLength: 1
          maxLength: 4 
          description: Codice dell'emoji usata come reazione
          pattern: '^[\u0000-\uFFFF]{1,4}$'
          example: "👍🏻"
        created_at:
          type: string 
          format: date-time
          example: "2025-10-30T14:48:00.000Z"
          description: Data e ora di creazione della reazione

    # Serve anche uno schema di input per le reazioni
...
    ForwardMessageRequest:
      type: object
      required:
        - target_chat_ids
      description: Richiesta per inoltrare un messaggio a una o più chat
      properties:
        target_chat_ids:
          type: array
          items:
            type: string
            pattern: '^[a-zA-Z0-9_-]{1,64}$'
            minLength: 1
            maxLength: 64
          description: ID delle chat di destinazione
          example: ["chat789", "chat999"]
          minItems: 1
          maxItems: 20

    # Creo uno schema per gli errori generici
    Error:
      type: object
      description: Rappresenta un errore generico restituito dalle API
      properties:
        code:
          type: string
          minLength: 3
          maxLength: 64
          pattern: '^[A-Z_]{3,64}$'
          example: "NOT_FOUND"
          description: Codice dell'errore
        message:
          type: string
          minLength: 3
          maxLength: 64
          pattern: '^[\s\S]{3,64}$'
          example: "La risorsa richiesta non è stata trovata."
          description: Descrizione dettagliata dell'errore

  # RESPONSES

  # Risposte per gli errori
  responses:
    BadRequest:
      description: Richiesta non valida (400 Bad Request)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Risorsa non trovata (404 Not Found)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalServerError:
      description: Errore interno del server (500 Internal Server Error)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  # PARAMETERS
  parameters:
...
    Ordinamento:
      name: order
      in: query
      description: Criterio di ordinamento (ascendente o discendente).
      required: false
      schema:
        type: string
        enum:
          - asc
          - desc
        default: desc

  # SECURITY SCHEMES
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT